add install condition

+++ docker-glyphhanger.sh
@@ -13,4 +13,9 @@ then
     docker pull wopolow/glyphhanger
 fi
 
-docker container run --rm -v $(pwd):/hwd wopolow/glyphhanger bash && glyphhanger "$@"
+if ! [ -z "$1" ] && [ $1 != "install" ]
+then 
+    docker container run --rm -v $(pwd):/hwd wopolow/glyphhanger bash && glyphhanger "$@"
+else
+    echo "docker-glyphhanger: internal installation complete";
+fi
